Add header file, separate format feature to another file
