+++ Makefile
@@ -3,7 +3,7 @@
 ##				 - use command "make debug" to compile and execute with GDB  		##
 
 all:
-	g++ -o IP_mng_app mng_app.cpp
+	g++ -o IP_mng_app mng_app.cpp format.cpp mng_app.h
 	clear
 	./IP_mng_app
